third_party/blink/web_tests/fast/css/xml-lang-ignored-in-html-expected.txt

This tests that the attribute in no namespace with no prefix and with the literal localname "xml:lang" is ignored in HTML. See bug 67586.

PASS languageOfNode('m1') is "auto"
PASS languageOfNode('n1') is "\"ko\""
PASS languageOfNode('o1') is "\"ar\""
PASS languageOfNode('o2') is "\"ar\""
PASS successfullyParsed is true

TEST COMPLETE
